+++ scripts/constants.js
@@ -1,3 +1,11 @@
+/*!
+ * Macro Manager
+ * Copyright (c) 2022 https://github.com/brunocalado
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3.
+ */
+
 export const MODULE_ID = "macro-manager";
 export const TEMPLATE_BUILDER = `modules/${MODULE_ID}/templates/builder.hbs`;
 export const TEMPLATE_WINDOW = `modules/${MODULE_ID}/templates/window.hbs`;